+++ src/aics-image-viewer/shared/utils/test/urlParsing.test.ts
@@ -1,20 +1,17 @@
 import type { CameraState } from "@aics/vole-core";
 import { describe, expect, it } from "@jest/globals";
 
-import type { ChannelState, ViewerState } from "../../../state/types";
-import { getDefaultCameraState, getDefaultChannelState, getDefaultViewerState } from "../../constants";
-import { ImageType, RenderMode, ViewMode } from "../../enums";
-import type { ViewerChannelStateParams, ViewerStateParams } from "../urlParsing";
+import { parseKeyValueList } from "../../../state/deserialize";
 import {
-  CONTROL_POINTS_REGEX,
-  deserializeViewerChannelSetting,
-  deserializeViewerState,
-  LEGACY_CONTROL_POINTS_REGEX,
-  parseHexColorAsColorArray,
-  parseKeyValueList,
-  parseStringEnum,
-  parseStringFloat,
-  parseStringInt,
+  CUSTOM_VIEWER_STATE,
+  DEFAULT_VIEWER_STATE,
+  SERIALIZED_CUSTOM_VIEWER_STATE,
+  SERIALIZED_DEFAULT_VIEWER_STATE,
+} from "../../../state/test/test_data";
+import type { ChannelState, ViewerChannelStateParams, ViewerState } from "../../../state/types";
+import { getDefaultCameraState, getDefaultChannelState, getDefaultViewerState } from "../../constants";
+import { ViewMode } from "../../enums";
+import {
   parseViewerUrlParams,
   serializeCameraState,
   serializeViewerChannelSetting,
@@ -23,6 +20,7 @@ import {
 } from "../urlParsing";
 import type { ViewerChannelSetting } from "../viewerChannelSettings";
 
+// TODO this is duplicated in state/test/deserialize.test.ts. Don't let me PR this without deduping!
 const defaultSettings: ViewerChannelSetting = {
   match: 0,
   color: undefined,
@@ -33,194 +31,6 @@ const defaultSettings: ViewerChannelSetting = {
   colorizeEnabled: undefined,
   colorizeAlpha: undefined,
 };
-
-//// VALUE PARSING ///////////////////////////////////////
-
-describe("LEGACY_CONTROL_POINTS_REGEX", () => {
-  it("accepts single control points", () => {
-    const data = "1:0.5:ffffff";
-    expect(LEGACY_CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("accepts multiple control points", () => {
-    const data = "1:0.5:ff0000,128:0.7:ffff00,255:1:ff0000";
-    expect(LEGACY_CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("accepts negative numbers", () => {
-    const data = "-1:0.5:ff0000,-255:1:ff0000";
-    expect(LEGACY_CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("allows 'w' as a placeholder for color strings", () => {
-    const data = "1:0.5:1,255:1:1";
-    expect(LEGACY_CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-});
-
-describe("CONTROL_POINTS_REGEX", () => {
-  it("accepts single control points", () => {
-    const data = "1:0.5:ffffff";
-    expect(CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("accepts multiple control points", () => {
-    const data = "1:0.5:ff0000:128:0.7:ffff00:255:1:ff0000";
-    expect(CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("accepts negative numbers", () => {
-    const data = "-1:0.5:ff0000:-255:1:ff0000";
-    expect(CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-
-  it("allows empty color strings", () => {
-    const data = "1:0.5:1:255:1:1";
-    expect(CONTROL_POINTS_REGEX.test(data)).toBe(true);
-  });
-});
-
-describe("parseKeyValueList", () => {
-  it("returns expected key value pairs", () => {
-    const data = "key1:value1,key2:value2";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({ key1: "value1", key2: "value2" });
-  });
-
-  it("decodes encoded key and value string names", () => {
-    const data = "ampersand%26:ampersand%26,comma%2C:comma%2C,colon%3A:colon%3A";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({ "ampersand&": "ampersand&", "comma,": "comma,", "colon:": "colon:" });
-  });
-
-  it("ignores repeat colons in value string", () => {
-    const data = "key1:value1:extra:extra,key2:value2";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({ key1: "value1:extra:extra", key2: "value2" });
-  });
-
-  it("returns empty object for empty string", () => {
-    const data = "";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({});
-  });
-
-  it("parses example viewer channel settings", () => {
-    const data = "col:FF0000,clz:1,cza:0.5,isa:0.75,ven:1,sen:1,isv:128,lut:autoij:";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({
-      col: "FF0000",
-      clz: "1",
-      cza: "0.5",
-      isa: "0.75",
-      ven: "1",
-      sen: "1",
-      isv: "128",
-      lut: "autoij:",
-    });
-  });
-
-  it("removes trailing and leading whitespace", () => {
-    const data = " key1 : value1 , key2 : value2 , key3: value 3";
-    const result = parseKeyValueList(data);
-    expect(result).toEqual({ key1: "value1", key2: "value2", key3: "value 3" });
-  });
-});
-
-describe("parseHexColorAsColorArray", () => {
-  it("parses hex values", () => {
-    expect(parseHexColorAsColorArray("000000")).toEqual([0, 0, 0]);
-    expect(parseHexColorAsColorArray("FFFFFF")).toEqual([255, 255, 255]);
-    expect(parseHexColorAsColorArray("ff0000")).toEqual([255, 0, 0]);
-    expect(parseHexColorAsColorArray("123456")).toEqual([18, 52, 86]);
-    expect(parseHexColorAsColorArray("7890ab")).toEqual([120, 144, 171]);
-    expect(parseHexColorAsColorArray("cdefef")).toEqual([205, 239, 239]);
-    expect(parseHexColorAsColorArray("ABCDEF")).toEqual([171, 205, 239]);
-  });
-
-  it("ignores unrecognized formats", () => {
-    expect(parseHexColorAsColorArray("f")).toBeUndefined();
-    expect(parseHexColorAsColorArray("fff")).toBeUndefined();
-    expect(parseHexColorAsColorArray("fffffff")).toBeUndefined();
-    expect(parseHexColorAsColorArray("hjklmn")).toBeUndefined();
-  });
-
-  it("parses the '1' code as white", () => {
-    expect(parseHexColorAsColorArray("1")).toEqual([255, 255, 255]);
-  });
-});
-
-describe("parseStringEnum", () => {
-  enum TestEnum {
-    SOME_VALUE = "some_value",
-    ANOTHER_VALUE = "another_value",
-  }
-
-  it("recognizes enum values", () => {
-    expect(parseStringEnum("some_value", TestEnum)).toEqual(TestEnum.SOME_VALUE);
-    expect(parseStringEnum("another_value", TestEnum)).toEqual(TestEnum.ANOTHER_VALUE);
-  });
-
-  it("returns default values for unrecognized enum values", () => {
-    expect(parseStringEnum("unexpected", TestEnum, undefined)).toEqual(undefined);
-    expect(parseStringEnum("unexpected", TestEnum, TestEnum.SOME_VALUE)).toEqual(TestEnum.SOME_VALUE);
-  });
-});
-
-describe("parseStringInt", () => {
-  it("parses int and undefined values", () => {
-    expect(parseStringInt("0", 0, 1000)).toEqual(0);
-    expect(parseStringInt("1", 0, 1000)).toEqual(1);
-    expect(parseStringInt("100", 0, 1000)).toEqual(100);
-    expect(parseStringInt("255", 0, 1000)).toEqual(255);
-    expect(parseStringInt("-1", -1000, 1000)).toEqual(-1);
-  });
-
-  it("returns undefined for undefined and NaN values", () => {
-    expect(parseStringInt("bad", -1000, 1000)).toEqual(undefined);
-    expect(parseStringInt("NaN", -1000, 1000)).toEqual(undefined);
-    expect(parseStringInt(undefined, -1000, 1000)).toEqual(undefined);
-  });
-
-  it("casts float values to integers", () => {
-    expect(parseStringInt("0.5", 0, 1000)).toEqual(0);
-    expect(parseStringInt("99.5", 0, 1000)).toEqual(99);
-    expect(parseStringInt("-10.5", -1000, 1000)).toEqual(-10);
-  });
-
-  it("applies clamping", () => {
-    expect(parseStringInt("0", 0, 255)).toEqual(0);
-    expect(parseStringInt("-1", 0, 255)).toEqual(0);
-    expect(parseStringInt("128", 0, 255)).toEqual(128);
-    expect(parseStringInt("255", 0, 255)).toEqual(255);
-    expect(parseStringInt("256", 0, 255)).toEqual(255);
-  });
-});
-
-describe("parseStringFloat", () => {
-  it("parses strings to float values", () => {
-    expect(parseStringFloat("0", -1000, 1000)).toEqual(0);
-    expect(parseStringFloat("1.0", -1000, 1000)).toEqual(1.0);
-    expect(parseStringFloat("0.5", -1000, 1000)).toEqual(0.5);
-    expect(parseStringFloat("128.5", -1000, 1000)).toEqual(128.5);
-    expect(parseStringFloat("495.344", -1000, 1000)).toEqual(495.344);
-    expect(parseStringFloat("-1", -1000, 1000)).toEqual(-1);
-  });
-
-  it("returns undefined for NaN or undefined values", () => {
-    expect(parseStringFloat("bad", -1000, 1000)).toEqual(undefined);
-    expect(parseStringFloat("NaN", -1000, 1000)).toEqual(undefined);
-    expect(parseStringFloat(undefined, -1000, 1000)).toEqual(undefined);
-  });
-
-  it("applies clamping", () => {
-    expect(parseStringFloat("0.5", 0, 1.0)).toEqual(0.5);
-    expect(parseStringFloat("1.0", 0, 1.0)).toEqual(1.0);
-    expect(parseStringFloat("1.1", 0, 1.0)).toEqual(1.0);
-    expect(parseStringFloat("0", 0, 1.0)).toEqual(0);
-    expect(parseStringFloat("-0.1", 0, 1.0)).toEqual(0);
-  });
-});
 
 //// SERIALIZE STATES ///////////////////////
 
@@ -258,107 +68,6 @@ describe("Channel state serialization", () => {
     ram: "0:255",
     pin: "0",
   };
-
-  // Note that the serialization + deserialization are NOT direct inverses.
-  // Serializing converts a ChannelState object to a ViewerChannelSettingParams object,
-  // deserializing converts a ViewerChannelSettingParams object to a **ViewerChannelSetting** object.
-
-  //  ChannelStates are per-channel and are used internally, while ViewerChannelSettings are
-  // input into the viewer and support grouping/matching to multiple channels.
-  // TODO: refactor these to all be the same state?
-
-  describe("deserializeViewerChannelSetting", () => {
-    it("returns default settings for empty objects", () => {
-      const data = {};
-      const result = deserializeViewerChannelSetting(0, data);
-      expect(result).toEqual(defaultSettings);
-    });
-
-    it("ignores unexpected keys", () => {
-      const data = { badKey: "badValue", ven: "1", sen: "1" } as ViewerChannelStateParams;
-      const result = deserializeViewerChannelSetting(0, data);
-      expect(result).toEqual({ ...defaultSettings, enabled: true, surfaceEnabled: true });
-    });
-
-    it("parses settings correctly", () => {
-      const data = {
-        col: "FF0000",
-        clz: "1",
-        cza: "0.5",
-        isa: "0.75",
-        ven: "1",
-        sen: "1",
-        isv: "128",
-        lut: "0:255",
-      } as ViewerChannelStateParams;
-      expect(deserializeViewerChannelSetting(0, data)).toEqual({
-        match: 0,
-        color: "FF0000",
-        enabled: true,
-        surfaceEnabled: true,
-        isovalue: 128,
-        surfaceOpacity: 0.75,
-        colorizeEnabled: true,
-        colorizeAlpha: 0.5,
-        intensity: { lut: ["0", "255"] },
-      });
-    });
-
-    it("handles expected lut formatting", () => {
-      const luts = [
-        ["autoij:0", ["autoij", "0"]],
-        ["0:autoij", ["0", "autoij"]],
-        [":autoij", ["", "autoij"]],
-        ["autoij:", ["autoij", ""]],
-        ["0:255", ["0", "255"]],
-        ["0.5:1.0", ["0.5", "1.0"]],
-        ["0.50:1.00", ["0.50", "1.00"]],
-        ["m99:m100", ["m99", "m100"]],
-        ["p10:p90", ["p10", "p90"]],
-        ["p10: p90", ["p10", "p90"]], // handle spaces
-      ];
-      for (const [encodedLut, decodedLut] of luts) {
-        const data = { lut: encodedLut } as ViewerChannelStateParams;
-        const result = deserializeViewerChannelSetting(0, data);
-        expect(result.intensity?.lut).toEqual(decodedLut);
-      }
-    });
-
-    it("ignores unexpected lut formats", () => {
-      const luts = ["!:0", "0:9:93", "255", ""];
-      for (const lut of luts) {
-        const data = { lut } as ViewerChannelStateParams;
-        const result = deserializeViewerChannelSetting(0, data);
-        expect(result.lut).toBeUndefined();
-      }
-    });
-
-    it("handles hex color formats", () => {
-      const colors = ["000000", "FFFFFF", "ffffff", "012345", "6789AB", "CDEF01", "abcdef"];
-      for (const color of colors) {
-        const data = { col: color } as ViewerChannelStateParams;
-        const result = deserializeViewerChannelSetting(0, data);
-        expect(result.color).toEqual(color);
-      }
-    });
-
-    it("ignores bad color formats", () => {
-      const badColors = ["f", "ff00", "red", "rgb(255,0,0)"];
-      for (const color of badColors) {
-        const data = { col: color } as ViewerChannelStateParams;
-        const result = deserializeViewerChannelSetting(0, data);
-        expect(result.color).toBeUndefined();
-      }
-    });
-
-    it("ignores bad float data", () => {
-      const data = { cza: "NaN", isa: "bad", isv: "f8" } as ViewerChannelStateParams;
-      const result = deserializeViewerChannelSetting(0, data);
-      expect(result.colorizeAlpha).toBeUndefined();
-      expect(result.surfaceOpacity).toBeUndefined();
-      expect(result.isovalue).toBeUndefined();
-    });
-  });
 
   describe("serializeViewerChannelSetting", () => {
     it("serializes channel settings", () => {
@@ -402,173 +111,24 @@ describe("Channel state serialization", () => {
   });
 });
 
-describe("Viewer state", () => {
-  const DEFAULT_VIEWER_STATE: ViewerState = {
-    viewMode: ViewMode.threeD, // "XY", "XZ", "YZ"
-    renderMode: RenderMode.volumetric, // "pathtrace", "maxproject"
-    imageType: ImageType.segmentedCell,
-    showAxes: false,
-    showBoundingBox: false,
-    backgroundColor: [0, 0, 0],
-    boundingBoxColor: [255, 255, 255],
-    autorotate: false,
-    maskAlpha: 0,
-    brightness: 70,
-    density: 50,
-    levels: [35, 140, 255],
-    interpolationEnabled: true,
-    region: { x: [0, 1], y: [0, 1], z: [0, 1] },
-    slice: { x: 0.5, y: 0.5, z: 0.5 },
-    time: 0,
-    scene: 0,
-    cameraState: undefined,
-    singleChannelMode: false,
-    singleChannelIndex: 0,
-    useExactScaleLevel: false,
-    scaleLevelIndex: 0,
-  };
-  const SERIALIZED_DEFAULT_VIEWER_STATE: ViewerStateParams = {
-    mode: "volumetric",
-    view: "3D",
-    image: "cell",
-    axes: "0",
-    bb: "0",
-    bgcol: "000000",
-    bbcol: "ffffff",
-    rot: "0",
-    mask: "0",
-    bright: "70",
-    dens: "50",
-    lvl: "35,140,255",
-    interp: "1",
-    reg: "0:1,0:1,0:1",
-    slice: "0.5,0.5,0.5",
-    t: "0",
-    scene: "0",
-    scm: "0",
-    sci: "0",
-    esl: "0",
-    scl: "0",
-  };
-
-  const CUSTOM_VIEWER_STATE: ViewerState = {
-    renderMode: RenderMode.pathTrace,
-    viewMode: ViewMode.xy,
-    imageType: ImageType.fullField,
-    showAxes: true,
-    showBoundingBox: true,
-    backgroundColor: [255, 0, 0],
-    boundingBoxColor: [0, 255, 0],
-    autorotate: true,
-    maskAlpha: 55,
-    brightness: 100,
-    density: 75,
-    levels: [0, 250, 251],
-    interpolationEnabled: false,
-    region: { x: [0, 0.5], y: [0, 1], z: [0, 1] },
-    slice: { x: 0.25, y: 0.75, z: 0.5 },
-    time: 100,
-    scene: 3,
-    singleChannelMode: true,
-    singleChannelIndex: 3,
-    useExactScaleLevel: true,
-    scaleLevelIndex: 3,
-    cameraState: {
-      position: [-1.05, -4, 45],
-      target: [0, 0, 0],
-      up: [0, 1, 0],
-      orthoScale: 3.534,
-      fov: 43.5,
-    },
-  };
-  const SERIALIZED_CUSTOM_VIEWER_STATE: ViewerStateParams = {
-    mode: "pathtrace",
-    view: "Z",
-    image: "fov",
-    axes: "1",
-    bb: "1",
-    bgcol: "ff0000",
-    bbcol: "00ff00",
-    rot: "1",
-    mask: "55",
-    bright: "100",
-    dens: "75",
-    lvl: "0,250,251",
-    interp: "0",
-    reg: "0:0.5,0:1,0:1",
-    slice: "0.25,0.75,0.5",
-    t: "100",
-    scene: "3",
-    scm: "1",
-    sci: "3",
-    esl: "1",
-    scl: "3",
-    cam: "pos:-1.05:-4:45,tar:0:0:0,up:0:1:0,ort:3.534,fov:43.5",
-  };
-
-  describe("serializeViewerState", () => {
-    it("serializes the default viewer settings", () => {
-      expect(serializeViewerState(DEFAULT_VIEWER_STATE, false)).toEqual(SERIALIZED_DEFAULT_VIEWER_STATE);
-    });
-
-    it("serializes custom viewer settings", () => {
-      expect(serializeViewerState(CUSTOM_VIEWER_STATE, false)).toEqual(SERIALIZED_CUSTOM_VIEWER_STATE);
-    });
-
-    it("shortens long numbers in the slice and region parameters", () => {
-      // Floats should be rounded to 7 significant digits or less
-      let state: Partial<ViewerState> = {
-        region: { x: [0.4566666666, 0.8667332], y: [0.49999999, 0.8999999], z: [0.3000000001, 0.16467883] },
-        slice: { x: 0.41111186, y: 0.49999999, z: 0.677402 },
-      };
-      let serializedState = serializeViewerState(state, true);
-      expect(serializedState.reg).toEqual("0.4566667:0.8667332,0.5:0.8999999,0.3:0.1646788");
-      expect(serializedState.slice).toEqual("0.4111119,0.5,0.677402");
-    });
+describe("serializeViewerState", () => {
+  it("serializes the default viewer settings", () => {
+    expect(serializeViewerState(DEFAULT_VIEWER_STATE, false)).toEqual(SERIALIZED_DEFAULT_VIEWER_STATE);
   });
 
-  describe("deserializeViewerState", () => {
-    it("returns an empty object for empty input", () => {
-      expect(deserializeViewerState({})).toEqual({});
-    });
+  it("serializes custom viewer settings", () => {
+    expect(serializeViewerState(CUSTOM_VIEWER_STATE, false)).toEqual(SERIALIZED_CUSTOM_VIEWER_STATE);
+  });
 
-    it("deserializes the default viewer settings", () => {
-      const params = SERIALIZED_DEFAULT_VIEWER_STATE;
-      expect(deserializeViewerState(params)).toEqual(DEFAULT_VIEWER_STATE);
-    });
-
-    it("deserializes custom viewer settings", () => {
-      const params = SERIALIZED_CUSTOM_VIEWER_STATE;
-      expect(deserializeViewerState(params)).toEqual(CUSTOM_VIEWER_STATE);
-    });
-
-    it("handles all ViewMode values", () => {
-      const viewModes = Object.values(ViewMode);
-      for (const viewMode of viewModes) {
-        const state: ViewerState = { ...DEFAULT_VIEWER_STATE, viewMode };
-        expect(deserializeViewerState(serializeViewerState(state, false)).viewMode).toEqual(viewMode);
-      }
-    });
-
-    it("handles all RenderMode values", () => {
-      const renderModes = Object.values(RenderMode);
-      for (const renderMode of renderModes) {
-        const state: ViewerState = { ...DEFAULT_VIEWER_STATE, renderMode };
-        expect(deserializeViewerState(serializeViewerState(state, false)).renderMode).toEqual(renderMode);
-      }
-    });
-
-    it("deserializes partial camera settings", () => {
-      const state: Partial<ViewerState> = {
-        cameraState: {
-          position: [1.0, -1.4, 45],
-          up: [0, 1, 0],
-          fov: 43.5,
-        },
-      };
-      const serializedState = "pos:1:-1.4:45,up:0:1:0,fov:43.5";
-      expect(deserializeViewerState({ cam: serializedState })).toEqual(state);
-    });
+  it("shortens long numbers in the slice and region parameters", () => {
+    // Floats should be rounded to 7 significant digits or less
+    let state: Partial<ViewerState> = {
+      region: { x: [0.4566666666, 0.8667332], y: [0.49999999, 0.8999999], z: [0.3000000001, 0.16467883] },
+      slice: { x: 0.41111186, y: 0.49999999, z: 0.677402 },
+    };
+    let serializedState = serializeViewerState(state, true);
+    expect(serializedState.reg).toEqual("0.4566667:0.8667332,0.5:0.8999999,0.3:0.1646788");
+    expect(serializedState.slice).toEqual("0.4111119,0.5,0.677402");
   });
 });
 
@@ -600,65 +160,7 @@ describe("Camera state", () => {
 
 //// DESERIALIZE STATES ///////////////////////
 
-describe("Channel state deserialization", () => {
-  const DEFAULT_CONTROL_POINTS = [
-    { x: -10, opacity: 0, color: [0, 0, 0] },
-    { x: 50, opacity: 0, color: [0, 0, 0] },
-    { x: 100, opacity: 0.3, color: [0, 16, 255] },
-    { x: 140, opacity: 0.8, color: [0, 255, 255] },
-    { x: 260, opacity: 1, color: [0, 255, 180] },
-  ];
-
-  describe("legacy control points", () => {
-    it("parses comma-separated control points", () => {
-      const result = deserializeViewerChannelSetting(0, {
-        cps: "-10:0:000000,50:0:000000,100:0.3:0010ff,140:0.8:00ffff,260:1:00ffb4",
-      });
-      expect(result.controlPoints).toEqual(DEFAULT_CONTROL_POINTS);
-    });
-
-    it("parses colon-separated control points", () => {
-      const result = deserializeViewerChannelSetting(0, {
-        cps: "-10:0:000000:50:0:000000:100:0.3:0010ff:140:0.8:00ffff:260:1:00ffb4",
-      });
-      expect(result.controlPoints).toEqual(DEFAULT_CONTROL_POINTS);
-    });
-
-    it("replaces '1' color strings with default color #ffffff", () => {
-      const result = deserializeViewerChannelSetting(0, {
-        cps: "0:0:1:50:1:1",
-      });
-      expect(result.controlPoints).toEqual([
-        { x: 0, opacity: 0, color: [255, 255, 255] },
-        { x: 50, opacity: 1, color: [255, 255, 255] },
-      ]);
-    });
-  });
-
-  it("parses comma-separated control points", () => {
-    const result = deserializeViewerChannelSetting(0, {
-      cpt: "-10:0:000000,50:0:000000,100:0.3:0010ff,140:0.8:00ffff,260:1:00ffb4",
-    });
-    expect(result.intensity?.controlPoints).toEqual(DEFAULT_CONTROL_POINTS);
-  });
-
-  it("parses colon-separated control points", () => {
-    const result = deserializeViewerChannelSetting(0, {
-      cpt: "-10:0:000000:50:0:000000:100:0.3:0010ff:140:0.8:00ffff:260:1:00ffb4",
-    });
-    expect(result.intensity?.controlPoints).toEqual(DEFAULT_CONTROL_POINTS);
-  });
-
-  it("replaces '1' color strings with default color #ffffff", () => {
-    const result = deserializeViewerChannelSetting(0, {
-      cpt: "0:0:1:50:1:1",
-    });
-    expect(result.intensity?.controlPoints).toEqual([
-      { x: 0, opacity: 0, color: [255, 255, 255] },
-      { x: 50, opacity: 1, color: [255, 255, 255] },
-    ]);
-  });
-});
+describe("Channel state deserialization", () => {});
 
 //// URL parsing /////////////////////////////////
