+++ src/aics-image-viewer/state/test/test_data.ts
@@ -1,0 +1,107 @@
+import { ImageType, RenderMode, ViewMode } from "../../shared/enums";
+import type { ViewerState, ViewerStateParams } from "../types";
+
+export const DEFAULT_VIEWER_STATE: ViewerState = {
+  viewMode: ViewMode.threeD, // "XY", "XZ", "YZ"
+  renderMode: RenderMode.volumetric, // "pathtrace", "maxproject"
+  imageType: ImageType.segmentedCell,
+  showAxes: false,
+  showBoundingBox: false,
+  backgroundColor: [0, 0, 0],
+  boundingBoxColor: [255, 255, 255],
+  autorotate: false,
+  maskAlpha: 0,
+  brightness: 70,
+  density: 50,
+  levels: [35, 140, 255],
+  interpolationEnabled: true,
+  region: { x: [0, 1], y: [0, 1], z: [0, 1] },
+  slice: { x: 0.5, y: 0.5, z: 0.5 },
+  time: 0,
+  scene: 0,
+  cameraState: undefined,
+  singleChannelMode: false,
+  singleChannelIndex: 0,
+  useExactScaleLevel: false,
+  scaleLevelIndex: 0,
+};
+
+export const SERIALIZED_DEFAULT_VIEWER_STATE: ViewerStateParams = {
+  mode: "volumetric",
+  view: "3D",
+  image: "cell",
+  axes: "0",
+  bb: "0",
+  bgcol: "000000",
+  bbcol: "ffffff",
+  rot: "0",
+  mask: "0",
+  bright: "70",
+  dens: "50",
+  lvl: "35,140,255",
+  interp: "1",
+  reg: "0:1,0:1,0:1",
+  slice: "0.5,0.5,0.5",
+  t: "0",
+  scene: "0",
+  scm: "0",
+  sci: "0",
+  esl: "0",
+  scl: "0",
+};
+
+export const CUSTOM_VIEWER_STATE: ViewerState = {
+  renderMode: RenderMode.pathTrace,
+  viewMode: ViewMode.xy,
+  imageType: ImageType.fullField,
+  showAxes: true,
+  showBoundingBox: true,
+  backgroundColor: [255, 0, 0],
+  boundingBoxColor: [0, 255, 0],
+  autorotate: true,
+  maskAlpha: 55,
+  brightness: 100,
+  density: 75,
+  levels: [0, 250, 251],
+  interpolationEnabled: false,
+  region: { x: [0, 0.5], y: [0, 1], z: [0, 1] },
+  slice: { x: 0.25, y: 0.75, z: 0.5 },
+  time: 100,
+  scene: 3,
+  singleChannelMode: true,
+  singleChannelIndex: 3,
+  useExactScaleLevel: true,
+  scaleLevelIndex: 3,
+  cameraState: {
+    position: [-1.05, -4, 45],
+    target: [0, 0, 0],
+    up: [0, 1, 0],
+    orthoScale: 3.534,
+    fov: 43.5,
+  },
+};
+
+export const SERIALIZED_CUSTOM_VIEWER_STATE: ViewerStateParams = {
+  mode: "pathtrace",
+  view: "Z",
+  image: "fov",
+  axes: "1",
+  bb: "1",
+  bgcol: "ff0000",
+  bbcol: "00ff00",
+  rot: "1",
+  mask: "55",
+  bright: "100",
+  dens: "75",
+  lvl: "0,250,251",
+  interp: "0",
+  reg: "0:0.5,0:1,0:1",
+  slice: "0.25,0.75,0.5",
+  t: "100",
+  scene: "3",
+  scm: "1",
+  sci: "3",
+  esl: "1",
+  scl: "3",
+  cam: "pos:-1.05:-4:45,tar:0:0:0,up:0:1:0,ort:3.534,fov:43.5",
+};
